+++ release.py
@@ -4,8 +4,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
+import re
 import subprocess
 import sys
+import time
 from pathlib import Path
 
 from aiosmtpd import __version__ as version
@@ -24,10 +26,8 @@ try:
 except FileNotFoundError:
     print("Please install 'twine' first")
     sys.exit(1)
-result = subprocess.run(
-    ["twine", "verify-upload"], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-)
-if result.returncode == 2:
+result = subprocess.run(["pip", "freeze"], stdout=subprocess.PIPE)
+if b"\ntwine-verify-upload==" not in result.stdout:
     print("*** Package twine-verify-upload is not yet installed.")
     print("*** Consider installing it. It is very useful :)")
     has_verify = False
@@ -45,6 +45,19 @@ GPG_SIGNING_ID = {GPG_SIGNING_ID or 'None'}
 choice = input(f"Release aiosmtpd {version} - correct? [y/N]: ")
 if choice.lower() not in ("y", "yes"):
     sys.exit("Release aborted")
+
+newsfile = Path(".") / "aiosmtpd" / "docs" / "NEWS.rst"
+with newsfile.open("rt") as fin:
+    want = re.compile("^" + re.escape(version) + r"\s*\(\d{4}-\d\d-\d\d\)")
+    for ln in fin:
+        m = want.match(ln)
+        if not m:
+            continue
+        break
+    else:
+        print(f"ERROR: I found no datestamped entry for {version} in NEWS.rst!")
+        sys.exit(1)
+
 if not GPG_SIGNING_ID:
     choice = input("You did not specify GPG signing ID! Continue? [y/N]: ")
     if choice.lower() not in ("y", "yes"):
@@ -75,6 +88,11 @@ try:
     subprocess.run(twine_up, check=True)
 
     if has_verify:
+        print("Waiting for package to be received by PyPI...", end="")
+        for i in range(10, 0, -1):
+            print(i, end=" ")
+            time.sleep(1.0)
+        print()
         twine_verif = ["twine", "verify_upload"] + DISTFILES
         subprocess.run(twine_verif, check=True)
 except subprocess.CalledProcessError as e: